refactor(lib): extract exec_command into tools/ (#1191)
